modularized the profile page
